% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{run_dge}
\alias{run_dge}
\title{Here using limma for #1.}
\usage{
run_dge(ge_object)
}
\arguments{
\item{`technology`}{needs to be a vector of "microarray" or "RNA-seq" values corresponding to the technologies in the corresponding members of `ge_object`. If left at NULL, all datasets in `ge_object` are assumed to be from microarrays.

*Could be passing in a concrete contrast argument instead with a more
complex contrast once needed (when/if implementing, remember that these
will need to be supplied with `make.names`'d variable names) - for now
working with simple contrasts, so don't need to implement this.
*This expression should involve all levels from the "active_levels" and
"reference_levels" lists.

Remember to pre-merge any synonymous levels before using OMA!}
}
\description{
Here using limma for #1.
}
