% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{run_ma}
\alias{run_ma}
\title{`run_ma` uses metafor's `rma` and `rma.mv` for meta-analysis.}
\usage{
run_ma(
  dat_results,
  dat_levels = NULL,
  id_var = "gene",
  es_var = "coeff",
  variance_var = NULL,
  se_var = NULL,
  p_adj_method = c("BH", p.adjust.methods),
  Q_p_cutoff = 0.05,
  inclusion_cutoff = 1,
  parallel = FALSE,
  n_cores = future::availableCores() - 1
)
}
\arguments{
\item{`dat_results`}{must be a data frame with, at the minimum, a "gene"
column, an effect size column, and a variance column.}

\item{`dat_levels`}{needs to be passed if we're dealing with multilevel
data. If this is the case, the level has to be specified for each data-
set. This will trigger the calculation of multi-level fixed and random
effects.}

\item{`es_var`}{here is unrestricted, but if the results are from `run_dge`,
then this should be either "cohen_d", or "coeff"; "coeff" is favored here,
as using "cohen_d" provides rescaling that also rescales the variance,
which is not accounted for by this method.}

\item{`variance_var`}{here is unrestricted, but if the results are from
`run_dge`, then this should be "p_pvalue".}

\item{`se_var`}{supply the name of the column containing the standard
errors if you've standard errors at your disposal rather than variances.}

\item{`parallel`}{- if set to TRUE, furrr will be used instead of purrr in
the M-A loop}
}
\description{
`run_ma` uses metafor's `rma` and `rma.mv` for meta-analysis.
}
